[#157]例外処理をまとめる
