+++ lib/entry_point_2018/exceptions/error_utility.rb
@@ -1,0 +1,10 @@
+module EntryPoint2018
+  class ErrorUtility
+    def self.log_and_notify(error)
+      Rails.logger.error error.class
+      Rails.logger.error error.message
+      Rails.logger.error error.backtrace.join("\n")
+      # Bugsnag.notify error
+    end
+  end
+end
